+++ src/stores/sign.store.ts
@@ -1,6 +1,7 @@
 import { atom, DefaultValue, RecoilState, selector } from 'recoil';
 
 import { SignUpType } from '@/types/signUp.type';
+import { EmailAuthType, UserType } from '@/types/auth.type';
 
 // ? type 집 유형 0: 원룸/오피스텔, 1: 빌라/연립, 2: 아파트, 3: 단독주택 @number
 export const SignUpProfileTypeAtom = atom<SignUpType['type']>({
@@ -123,3 +124,60 @@ export const SignUpProfileSelector = selector<SignUpType>({
     }
   },
 });
+
+export const ShowVerificationAtom = atom({
+	key: 'showVerificationAtom',
+	default: false,
+})
+
+export const SignUpUserNameAtom = atom<UserType['name']>({
+  key: 'signUpUserNameAtom',
+  default: '',
+});
+
+export const SignUpUserBirthAtom = atom<UserType['birth']>({
+  key: 'signUpUserBirthAtom',
+  default: undefined,
+});
+
+export const SignUpUserGenderAtom = atom<UserType['gender']>({
+  key: 'signUpUserGenderAtom',
+  default: undefined,
+});
+
+export const SignUpUserEmailAtom = atom<EmailAuthType['email']>({
+  key: 'signUpUserEmailAtom',
+  default: undefined,
+});
+
+export const SignUpUserPasswordAtom = atom<EmailAuthType['password']>({
+  key: 'signUpUserPasswordAtom',
+  default: undefined,
+});
+
+export const SignUpTokenAtom = atom<EmailAuthType['token']>({
+  key: 'signUpTokenAton',
+  default: undefined,
+});
+
+export const SignUpIntroSelector = selector<UserType & EmailAuthType>(
+  {
+    key: 'SignUpIntroSelector',
+    get: ({ get }) => {
+      const name = get(SignUpUserNameAtom);
+      const birth = get(SignUpUserBirthAtom);
+      const gender = get(SignUpUserGenderAtom);
+      const email = get(SignUpUserEmailAtom);
+      const password = get(SignUpUserPasswordAtom);
+      const token = get(SignUpTokenAtom);
+      return {
+        name,
+        birth,
+        gender,
+        email,
+        password,
+        token,
+      };
+    },
+  },
+);
